(feat) move paper_trade to new location
