+++ hummingbot/connector/exchange/paper_trade/__init__.py
@@ -5,8 +5,8 @@ from hummingbot.connector.exchange.binance.binance_order_book_tracker import Bin
 from hummingbot.connector.exchange.coinbase_pro.coinbase_pro_order_book_tracker import CoinbaseProOrderBookTracker
 from hummingbot.market.huobi.huobi_market import HuobiMarket
 from hummingbot.market.huobi.huobi_order_book_tracker import HuobiOrderBookTracker
-from hummingbot.market.paper_trade.market_config import MarketConfig
-from hummingbot.market.paper_trade.paper_trade_market import PaperTradeMarket
+from hummingbot.connector.exchange.paper_trade.market_config import MarketConfig
+from hummingbot.connector.exchange.paper_trade.paper_trade_market import PaperTradeMarket
 from hummingbot.connector.exchange.radar_relay.radar_relay_market import RadarRelayMarket
 from hummingbot.connector.exchange.radar_relay.radar_relay_order_book_tracker import RadarRelayOrderBookTracker
 from hummingbot.market.dolomite.dolomite_order_book_tracker import DolomiteOrderBookTracker
