+++ hummingbot/connector/exchange/paper_trade/paper_trade_market.pxd
