+++ hummingbot/connector/exchange/paper_trade/trading_pair.py
