+++ test/integration/test_paper_trade_market.py
@@ -31,9 +31,9 @@ from hummingbot.core.utils.async_utils import (
 )
 from hummingbot.core.event.event_logger import EventLogger
 from hummingbot.connector.exchange.binance.binance_order_book_tracker import BinanceOrderBookTracker
-from hummingbot.market.paper_trade.paper_trade_market import PaperTradeMarket, QueuedOrder
-from hummingbot.market.paper_trade.trading_pair import TradingPair
-from hummingbot.market.paper_trade.market_config import MarketConfig
+from hummingbot.connector.exchange.paper_trade.paper_trade_market import PaperTradeMarket, QueuedOrder
+from hummingbot.connector.exchange.paper_trade.trading_pair import TradingPair
+from hummingbot.connector.exchange.paper_trade.market_config import MarketConfig
 import pandas as pd
 from typing import List, Iterator, NamedTuple, Dict
 import logging
